% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{apply_secondary_models}
\alias{apply_secondary_models}
\title{Calculates the effect of the secondary models on the model parameters for some environmental conditions}
\usage{
apply_secondary_models(env_conditions, secondary_models)
}
\arguments{
\item{env_conditions}{A list with as many entries as environmental factors. Each is
a numeric vector of length 1 (a tibble of 1 row also works).}

\item{secondary_models}{A nested list describing the secondary models (as in \code{\link[=predict_inactivation]{predict_inactivation()}}).}
}
\value{
A list, with as many entries as model parameters, with the parameter values.
}
\description{
Calculates the effect of the secondary models on the model parameters for some environmental conditions
}
