% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metainfo_dynamic.R
\name{check_dynamic_pars}
\alias{check_dynamic_pars}
\title{Basic check of parameters for predictions under dynamic conditions}
\usage{
check_dynamic_pars(primary_model, secondary_models)
}
\arguments{
\item{primary_model}{A named list as in \code{\link[=predict_inactivation]{predict_inactivation()}}}

\item{secondary_models}{A named list as in \code{\link[=predict_inactivation]{predict_inactivation()}}}
}
\value{
If there is no error, the model function.
}
\description{
Checks that: AA
}
