% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_guess.R
\name{check_secondary_guess}
\alias{check_secondary_guess}
\title{Checking guesses for secondary inactivation models}
\usage{
check_secondary_guess(fit_data, model_name, guess, formula = my_par ~ temp)
}
\arguments{
\item{fit_data}{a tibble (or data.frame) with the data to use for the fit, as in
\code{\link[=fit_inactivation_secondary]{fit_inactivation_secondary()}}}

\item{model_name}{a secondary model name as in \code{\link[=secondary_model_data]{secondary_model_data()}}}

\item{guess}{a named numeric vector with the model parameters}

\item{formula}{a formula defining the model variables as in \code{\link[=fit_inactivation_secondary]{fit_inactivation_secondary()}}}
}
\value{
an instance of ggplot with two subplots: an observed vs predicted plot
and a predicted vs residuals plot
}
\description{
Checking guesses for secondary inactivation models
}
