% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metainfo_secondary.R
\name{check_secondary_pars}
\alias{check_secondary_pars}
\title{Basic check of parameters for fitting secondary models}
\usage{
check_secondary_pars(model_name, pars, vars)
}
\arguments{
\item{model_name}{Model identifier}

\item{pars}{A named list of model parameters}
}
\value{
If there is no error, the model function.
}
\description{
Checks that: the model name is correct, the right number of model
parameters have been defined and that the parameters have the right names
}
