% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{compare_inactivation_fits}
\alias{compare_inactivation_fits}
\title{Model comparison and selection for inactivation models}
\usage{
compare_inactivation_fits(models)
}
\arguments{
\item{models}{a (we recommend named) list of models fitted using \code{\link[=fit_inactivation]{fit_inactivation()}}.
Every model should be of the same class. Otherwise, some functions may give unexpected results.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function is a constructor for \link{InactivationComparison} or \link{GlobalComparison},
a class that provides several functions for model comparison and model selection
for inactivation models fitted using \code{\link[=fit_inactivation]{fit_inactivation()}}. Please see the help pages for
\link{InactivationComparison} or \link{GlobalComparison} for further details.

Although it is not necessary, we recommend passing the models as a named list,
as these names will later be kept in plots and tables.
}
