% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_dynamic_guess}
\alias{convert_dynamic_guess}
\title{A helper to convert guesses to the format for predictions}
\usage{
convert_dynamic_guess(sec_models, guess, known)
}
\description{
A helper to convert guesses to the format for predictions
}
