% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_models_dynamic.R
\name{dyna_acclimation}
\alias{dyna_acclimation}
\title{Modification of the Bigelow model with stress aclimation}
\usage{
dyna_acclimation(time, state, primary_pars, env_interpolator, secondary_models)
}
\arguments{
\item{time}{a 1D numeric with the treatment time}

\item{state}{a named numeric vector with one element defining the microbial concentration}

\item{primary_pars}{ignored. Just a dumpster for the param argument from ode}

\item{env_interpolator}{an interpolator for the environmental conditions as returned by \code{\link[=approx_env]{approx_env()}}}

\item{secondary_models}{a nested list defining the secondary models as per \code{\link[=apply_secondary_models]{apply_secondary_models()}}}
}
\value{
a list with dP/dt dN/dt as per \code{\link[=ode]{ode()}}
}
\description{
The functions is defined to be called from \code{\link[=ode]{ode()}} from the deSolve package.
}
