% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_models_dynamic.R
\name{dyna_Peleg}
\alias{dyna_Peleg}
\title{Peleg model for predictions under dynamic conditions}
\usage{
dyna_Peleg(
  time,
  state,
  primary_pars,
  env_interpolator,
  secondary_models,
  tol_logS = 1e-12
)
}
\arguments{
\item{time}{a 1D numeric with the treatment time}

\item{state}{a named numeric vector with one element defining the microbial concentration}

\item{primary_pars}{ignored. Just a dumpster for the param argument from ode}

\item{env_interpolator}{an interpolator for the environmental conditions as returned by \code{\link[=approx_env]{approx_env()}}}

\item{secondary_models}{a nested list defining the secondary models as per \code{\link[=apply_secondary_models]{apply_secondary_models()}}}

\item{tol_time}{a tolerance for the time (to avoid singularities for t=0). See Details}
}
\value{
a list with dN/dt as per \code{\link[=ode]{ode()}}
}
\description{
The functions is defined to be called from \code{\link[=ode]{ode()}} from the deSolve package.
}
\details{
The Peleg model under dynamic conditions can be singular for t=0 for some combinations
of model parameters. This is fixed by adding a tolerance to every time point. By default,
the tolerance is 1e-12, so it should not affect the calculations. Nonetheless,
this can be changed using the \code{tol_time} argument.
}
