% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dynamic_conditions}
\alias{dynamic_conditions}
\title{Dynamic treatment conditions}
\format{
A tibble with 3 rows and 3 columns:
\describe{
\item{time}{elapsed time in min}
\item{temp}{treatment temperature (C)}
\item{pH}{treatment pH}
}
}
\usage{
dynamic_conditions
}
\description{
An example dataset to illustrate \code{\link[=predict_inactivation]{predict_inactivation()}} and \code{\link[=fit_inactivation]{fit_inactivation()}}
under dynamic conditions. It describes a dummy dynamic treatment dependent on
temperature and pH
}
\keyword{datasets}
