% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dynamic_experiment}
\alias{dynamic_experiment}
\title{Results of a dynamic treatment}
\format{
A tibble with 20 rows and 3 columns:
\describe{
\item{time}{elapsed time in min}
\item{logN}{observed microbial concentration (log CFU/mL)}
}
}
\usage{
dynamic_experiment
}
\description{
An example dataset to illustrate \code{\link[=fit_inactivation]{fit_inactivation()}}
under dynamic conditions. It describes simulated data observed under the
dynamic treatment \link{dynamic_conditions}.
}
\keyword{datasets}
