% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_fit.R
\name{dynamic_residuals}
\alias{dynamic_residuals}
\title{Residuals for dynamic fitting}
\usage{
dynamic_residuals(
  this_p,
  fit_data,
  primary_model_name,
  sec_models,
  known,
  env_conditions,
  cost = NULL
)
}
\description{
Residuals for dynamic fitting
}
