% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_fit.R
\name{fit_dynamic_inactivation}
\alias{fit_dynamic_inactivation}
\title{AA}
\usage{
fit_dynamic_inactivation(
  fit_data,
  model_name,
  start,
  known,
  upper,
  lower,
  secondary_models,
  algorithm,
  env_conditions,
  niter
)
}
\description{
AA
}
