% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_inactivation.R
\name{fit_inactivation}
\alias{fit_inactivation}
\title{Fits an inactivation model using different methods from predictive microbiology}
\usage{
fit_inactivation(
  method,
  fit_data,
  primary_model_name,
  guess,
  known,
  upper = NULL,
  lower = NULL,
  secondary_models = NULL,
  algorithm = "regression",
  env_conditions = NULL,
  niter = NULL,
  check = TRUE,
  formula = logN ~ time
)
}
\arguments{
\item{method}{method for model fitting. One of 'primary', 'two-steps', 'one-step',
'dynamic' or 'global'.}

\item{fit_data}{a tibble with the data to use for model fitting. See sections below for further information.}

\item{primary_model_name}{aa}

\item{guess}{description}

\item{known}{aa}

\item{upper}{aa}

\item{lower}{description}

\item{secondary_models}{description}

\item{algorithm}{description}

\item{niter}{description}

\item{check}{aa}

\item{env_conditinos}{description}
}
\description{
Fits an inactivation model using different methods from predictive microbiology
}
