% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inactivation_experiment}
\alias{inactivation_experiment}
\title{Results of an isothermal treatment}
\format{
A tibble with 11 rows and 3 columns:
\describe{
\item{time}{elapsed time in min}
\item{logN}{observed microbial concentration (log CFU/mL)}
}
}
\usage{
inactivation_experiment
}
\description{
An example dataset to illustrate \code{\link[=fit_inactivation]{fit_inactivation()}}
under isothermal conditions using simulated data
}
\keyword{datasets}
