% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inactivation_onestep}
\alias{inactivation_onestep}
\title{Data for one-step fitting}
\format{
A tibble with 25 rows and 4 columns:
\describe{
\item{time}{treatment time in min}
\item{logN}{observed microbial concentration (log CFU/mL)}
\item{temp}{treatment temperature (C)}
\item{pH}{treatment pH}
}
}
\usage{
inactivation_onestep
}
\description{
An example dataset to illustrate \code{\link[=fit_inactivation]{fit_inactivation()}}
using the one-step fitting approach. It simulates a batch of experiments at constant
environmental conditions changing two environmental factors.
}
\keyword{datasets}
