% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inactivation_secondary}
\alias{inactivation_secondary}
\title{Data for fitting a secondary inactivation model}
\format{
A tibble with 25 rows and 3 columns:
\describe{
\item{temp}{treatment temperature (C)}
\item{pH}{treatment pH}
\item{delta}{the estimate of delta (min)}
}
}
\usage{
inactivation_secondary
}
\description{
An example dataset to illustrate \code{\link[=fit_inactivation_secondary]{fit_inactivation_secondary()}}.
It simulates the value of \code{delta} estimated from a batch of experiments at
constant conditions changing two environmental factors.
}
\keyword{datasets}
