% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inactivation_twosteps}
\alias{inactivation_twosteps}
\title{Data for two-step fitting}
\format{
A data.frame with 135 rows and 4 columns:
\describe{
\item{temp}{treatment temperature (C)}
\item{pH}{treatment pH}
\item{time}{treatment time in min}
\item{logN}{observed microbial concentration (log CFU/mL)}
}
}
\usage{
inactivation_twosteps
}
\description{
An example dataset to illustrate \code{\link[=fit_inactivation]{fit_inactivation()}} using the two-steps approach.
It simulates the change in the microbial concentration observed in 9 independent experiments
at constant conditions changing two environmental factors.
}
\keyword{datasets}
