% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InactivationFit_class.R
\name{InactivationFit}
\alias{InactivationFit}
\alias{coef.InactivationFit}
\alias{summary.InactivationFit}
\alias{predict.InactivationFit}
\alias{residuals.InactivationFit}
\alias{vcov.InactivationFit}
\alias{deviance.InactivationFit}
\alias{fitted.InactivationFit}
\alias{logLik.InactivationFit}
\alias{AIC.InactivationFit}
\alias{plot.InactivationFit}
\title{InactivationFit class}
\usage{
\method{coef}{InactivationFit}(object, ..., step = 2)

\method{summary}{InactivationFit}(object, ..., step = 2)

\method{predict}{InactivationFit}(object, times = NULL, env_conditions = NULL, ...)

\method{residuals}{InactivationFit}(object, ...)

\method{vcov}{InactivationFit}(object, ...)

\method{deviance}{InactivationFit}(object, ...)

\method{fitted}{InactivationFit}(object, ...)

\method{logLik}{InactivationFit}(object, ...)

\method{AIC}{InactivationFit}(object, ..., k = 2)

\method{plot}{InactivationFit}(
  x,
  y = NULL,
  ...,
  add_factor = NULL,
  line_col = "black",
  line_size = 1,
  line_type = 1,
  point_col = "black",
  point_size = 3,
  point_shape = 16,
  ylims = NULL,
  label_y1 = NULL,
  label_y2 = add_factor,
  label_x = "time",
  line_col2 = "black",
  line_size2 = 1,
  line_type2 = "dashed",
  type = 1
)
}
\arguments{
\item{object}{an instance of InactivationFit}

\item{...}{ignored.}

\item{times}{numeric vector describing the time points for the prediction.
If \code{NULL} (default), uses the same points as those used for fitting.}

\item{env_conditions}{tibble describing the environmental conditions as in \code{\link[=fit_growth]{fit_growth()}}.
If \code{NULL} (default), uses the environmental condition of the fitting.}

\item{k}{penalty for the parameters (k=2 by default)}

\item{x}{The object of class \link{InactivationFit} to plot.}

\item{y}{ignored}

\item{add_factor}{whether to plot also one environmental factor.
If \code{NULL} (default), no environmental factor is plotted. If set
to one character string that matches one entry of x$env_conditions,
that condition is plotted in the secondary axis. Ignored if \code{environment="constant"}}

\item{line_col}{Aesthetic parameter to change the colour of the line geom in the plot, see: \code{\link[=geom_line]{geom_line()}}}

\item{line_size}{Aesthetic parameter to change the thickness of the line geom in the plot, see: \code{\link[=geom_line]{geom_line()}}}

\item{line_type}{Aesthetic parameter to change the type of the line geom in the plot, takes numbers (1-6) or strings ("solid") see: \code{\link[=geom_line]{geom_line()}}}

\item{point_col}{Aesthetic parameter to change the colour of the point geom, see: \code{\link[=geom_point]{geom_point()}}}

\item{point_size}{Aesthetic parameter to change the size of the point geom, see: \code{\link[=geom_point]{geom_point()}}}

\item{point_shape}{Aesthetic parameter to change the shape of the point geom, see: \code{\link[=geom_point]{geom_point()}}}

\item{ylims}{A two dimensional vector with the limits of the primary y-axis.
\code{NULL} by default}

\item{label_y1}{Label of the primary y-axis.}

\item{label_y2}{Label of the secondary y-axis. Ignored if \code{environment="constant"}}

\item{label_x}{Label of the x-axis}

\item{line_col2}{Same as lin_col, but for the environmental factor. Ignored if \code{environment="constant"}}

\item{line_size2}{Same as line_size, but for the environmental factor. Ignored if \code{environment="constant"}}

\item{line_type2}{Same as lin_type, but for the environmental factor. Ignored if \code{environment="constant"}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The \code{InactivationFit} class contains an inactivation model fitted to data under
static or dynamic conditions. Its constructor is \code{\link[=fit_inactivation]{fit_inactivation()}}.

It is a subclass of list with the items:
\itemize{
\item method: fitting method as in \code{\link[=fit_inactivation]{fit_inactivation()}}
\item algorithm: type of algorithm as in \code{\link[=fit_inactivation]{fit_inactivation()}}
\item data: data used for model fitting
\item guess: initial guess of the model parameters
\item known: fixed model parameters
\item primary_model: a character describing the primary model
\item fit_results: an instance of modFit or modMCMC with the results of the fit
\item best_prediction: Instance of \link{InactivationPrediction} with the best inactivation fit
\item sec_models: a named vector with the secondary models assigned for each
environmental factor. \code{NULL} for \code{environment="constant"}
\item env_conditions: a tibble with the environmental conditions used for model
fitting. \code{NULL} for \code{environment="constant"}
\item niter: number of iterations of the Markov chain. \code{NULL} if \code{algorithm != "MCMC"}
\item logbase_logN: base of the logarithm for the definition of the population size
(check the relevant vignette)
\item approach_logN0: Approach for the initial microbial count when fitting several experiments
}
}
\section{Methods (by generic)}{
\itemize{
\item \code{coef(InactivationFit)}: vector of fitted model parameters.

\item \code{summary(InactivationFit)}: statistical summary of the fit.

\item \code{predict(InactivationFit)}: vector of model predictions.

\item \code{residuals(InactivationFit)}: vector of model residuals.

\item \code{vcov(InactivationFit)}: variance-covariance matrix of the model, estimated
as 1/(0.5*Hessian) for regression and as the variance-covariance of the draws
for MCMC

\item \code{deviance(InactivationFit)}: deviance of the model.

\item \code{fitted(InactivationFit)}: vector of fitted values.

\item \code{logLik(InactivationFit)}: loglikelihood of the model

\item \code{AIC(InactivationFit)}: Akaike Information Criterion

\item \code{plot(InactivationFit)}: compares the fitted model against the data.

}}
