% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InactivationPrediction_class.R
\name{InactivationPrediction}
\alias{InactivationPrediction}
\alias{print.InactivationPrediction}
\alias{coef.InactivationPrediction}
\alias{plot.InactivationPrediction}
\alias{summary.InactivationPrediction}
\title{InactivationPrediction class}
\usage{
\method{print}{InactivationPrediction}(x, ...)

\method{coef}{InactivationPrediction}(object, ...)

\method{plot}{InactivationPrediction}(
  x,
  y = NULL,
  ...,
  add_factor = NULL,
  ylims = NULL,
  label_y1 = NULL,
  label_y2 = add_factor,
  line_col = "black",
  line_size = 1,
  line_type = "solid",
  line_col2 = "black",
  line_size2 = 1,
  line_type2 = "dashed",
  label_x = "time",
  type = 1
)

\method{summary}{InactivationPrediction}(object, ...)
}
\arguments{
\item{x}{The object of class \code{InactivationPrediction} to plot.}

\item{...}{ignored}

\item{object}{An instance of \code{InactivationPrediction}.}

\item{y}{ignored}

\item{add_factor}{whether to plot also one environmental factor.
If \code{NULL} (default), no environmental factor is plotted. If set
to one character string that matches one entry of x$env_conditions,
that condition is plotted in the secondary axis. Ignored for \code{environment="constant"}.}

\item{ylims}{A two dimensional vector with the limits of the primary y-axis.}

\item{label_y1}{Label of the primary y-axis.}

\item{label_y2}{Label of the secondary y-axis.}

\item{line_col}{Aesthetic parameter to change the colour of the line geom in the plot, see: \code{\link[=geom_line]{geom_line()}}}

\item{line_size}{Aesthetic parameter to change the thickness of the line geom in the plot, see: \code{\link[=geom_line]{geom_line()}}}

\item{line_type}{Aesthetic parameter to change the type of the line geom in the plot, takes numbers (1-6) or strings ("solid") see: \code{\link[=geom_line]{geom_line()}}}

\item{line_col2}{Same as lin_col, but for the environmental factor.}

\item{line_size2}{Same as line_size, but for the environmental factor.}

\item{line_type2}{Same as lin_type, but for the environmental factor.}

\item{label_x}{Label of the x-axis.}

\item{type}{type of plot. See details.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The \code{InactivationPrediction} class contains the results of an inactivation prediction.
Its constructor is \code{\link[=predict_inactivation]{predict_inactivation()}}.

It is a subclass of list with the items:
\itemize{
\item simulation: a tibble with the model simulation
\item primary model: a list describing the primary model as in \code{\link[=predict_inactivation]{predict_inactivation()}}
\item environment: a character describing the type of environmental conditions
as in \code{\link[=predict_inactivation]{predict_inactivation()}}
\item env_conditions: a named list with the functions used to approximate the (dynamic)
environmental conditions. \code{NULL} if \code{environment="constant"}.
\item secondary_models: a named list describing the secondary models as in \code{\link[=predict_inactivation]{predict_inactivation()}}.
\code{NULL} if \code{environment="constant"}.
\item sec_effects: a tibble describing the effect of each environmental factor per time point.
\item logbase_mu: the log-base for the definition of parameter mu (see the relevant vignette)
\item logbase_logN: the log-base for the definition of the logarithm of the population size
}
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(InactivationPrediction)}: print of the model

\item \code{coef(InactivationPrediction)}: coefficients of the model

\item \code{plot(InactivationPrediction)}: predicted inactivation curve.

\item \code{summary(InactivationPrediction)}: summary of the model

}}
