% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_models_constant.R
\name{iso_Bigelow}
\alias{iso_Bigelow}
\title{Bigelow (log-lineal) model for constant conditions}
\usage{
iso_Bigelow(times, N0, D)
}
\arguments{
\item{times}{a numerical vector of time points}

\item{N0}{initial microbial concentration (1D numeric)}

\item{D}{D-value (1D numeric)}
}
\value{
A numerical vector of log-microbial concentrations
}
\description{
Bigelow (log-lineal) model for constant conditions
}
