% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_models_constant.R
\name{iso_Geeraerd_k}
\alias{iso_Geeraerd_k}
\title{Geeraerd model for constant conditions (based on k)}
\usage{
iso_Geeraerd_k(times, N0, SL, k, Nres)
}
\arguments{
\item{times}{a numerical vector of time points}

\item{N0}{initial microbial concentration (1D numeric)}

\item{SL}{shoulder length (1D numeric)}

\item{k}{inactivation rate (1D numeric)}

\item{Nres}{microbial concentration at the tail (1D numeric)}
}
\value{
A numerical vector of log-microbial concentrations
}
\description{
Geeraerd model for constant conditions (based on k)
}
