% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_models_constant.R
\name{iso_Geeraerd_noTail}
\alias{iso_Geeraerd_noTail}
\title{Geeraerd model for constant conditions without tail (based on K)}
\usage{
iso_Geeraerd_noTail(times, N0, SL, D)
}
\arguments{
\item{times}{a numerical vector of time points}

\item{N0}{initial microbial concentration (1D numeric)}

\item{SL}{shoulder length (1D numeric)}

\item{D}{D-value (1D numeric)}
}
\value{
A numerical vector of log-microbial concentrations
}
\description{
Geeraerd model for constant conditions without tail (based on K)
}
