% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_models_constant.R
\name{iso_Mafart}
\alias{iso_Mafart}
\title{Mafart (Weibullian) model for constant conditions}
\usage{
iso_Mafart(times, N0, delta, p)
}
\arguments{
\item{times}{a numerical vector of time points}

\item{N0}{initial microbial concentration (1D numeric)}

\item{delta}{delta-value (1D numeric)}

\item{p}{shape factor of the Mafart model (1D numeric)}
}
\value{
A numerical vector of log-microbial concentrations
}
\description{
Mafart (Weibullian) model for constant conditions
}
