% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_models_constant.R
\name{iso_Metselaar}
\alias{iso_Metselaar}
\title{Metselaar (Weibull) model for constant conditions}
\usage{
iso_Metselaar(times, Delta, N0, D, p)
}
\arguments{
\item{times}{a numerical vector of time points}

\item{Delta}{Reference log-reductions for the model (1D numeric)}

\item{N0}{initial microbial concentration (1D numeric)}

\item{D}{Equivalent D-value for Delta log-reductions (1D numeric)}

\item{p}{shape factor of the Weibull model (1D numeric)}
}
\value{
A numerical vector of log-microbial concentrations
}
\description{
Metselaar (Weibull) model for constant conditions
}
