% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_models_constant.R
\name{iso_Peleg}
\alias{iso_Peleg}
\title{Peleg (Weibullian) model for constant conditions}
\usage{
iso_Peleg(times, N0, b, n)
}
\arguments{
\item{times}{a numerical vector of time points}

\item{N0}{initial microbial concentration (1D numeric)}

\item{b}{b-value of the Peleg model (1D numeric)}

\item{n}{shape factor of the Peleg model (1D numeric)}
}
\value{
A numerical vector of log-microbial concentrations
}
\description{
Peleg (Weibullian) model for constant conditions
}
