% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_models_constant.R
\name{iso_Trilinear}
\alias{iso_Trilinear}
\title{Trilineal inactivatino model}
\usage{
iso_Trilinear(times, N0, SL, D, Nres)
}
\arguments{
\item{times}{a numerical vector of time points}

\item{N0}{initial microbial concentration (1D numeric)}

\item{SL}{shoulder length (1D numeric)}

\item{D}{D-value (1D numeric)}

\item{Nres}{microbial concentration at the tail (1D numeric)}
}
\value{
A numerical vector of log-microbial concentrations
}
\description{
Trilineal inactivatino model
}
