% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_models_constant.R
\name{iso_Weibull_2phase}
\alias{iso_Weibull_2phase}
\title{Weibull model with two populations}
\usage{
iso_Weibull_2phase(times, N0, delta1, delta2, p1, p2, alpha)
}
\arguments{
\item{times}{a numerical vector of time points}

\item{N0}{initial microbial concentration (1D numeric)}

\item{delta1}{delta-value of the 1st population (1D numeric)}

\item{delta2}{delta-value of the 2nd population (1D numeric)}

\item{p1}{shape factor of the Weibull model for the 1st population (1D numeric)}

\item{p2}{shape factor of the Weibull model for the 2nd population (1D numeric)}

\item{alpha}{parameter describing the population ratio (1D numeric)}
}
\value{
A numerical vector of log-microbial concentrations
}
\description{
Weibull model with two populations
}
