% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_guesses.R
\name{make_guess_factor}
\alias{make_guess_factor}
\title{Guess for one environmental factor}
\usage{
make_guess_factor(model_name, x, y)
}
\arguments{
\item{model_name}{a 1D character vector defining the secondary model}

\item{x}{a numeric vector with the input variable (the environmental factor)}

\item{y}{a numeric vector with the output variable (the parameter values)}
}
\description{
Guess for one environmental factor
}
