% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_guesses.R
\name{make_guess_primary}
\alias{make_guess_primary}
\title{Initial guesses for fitting primary inactivation models}
\usage{
make_guess_primary(fit_data, primary_model, formula = logN ~ time)
}
\arguments{
\item{fit_data}{the experimental data. A tibble (or data.frame) with a column
named \code{time} with the elapsed time and one called \code{logN} with the logarithm
of the population size}

\item{primary_model}{a string defining the equation of the primary model,
as defined in \code{\link[=primary_model_data]{primary_model_data()}}}

\item{formula}{an object of class "formula" describing the x and y variables.
\code{logN ~ time} as a default.}
}
\value{
A named numeric vector of initial guesses for the model parameters
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The function uses some heuristics to provide initial guesses for the parameters
of the growth model selected that can be used with \code{\link[=fit_inactivation]{fit_inactivation()}}.
}
