% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPC_fitting.R
\name{most_probable_curve}
\alias{most_probable_curve}
\title{Most Probable Curve method}
\usage{
most_probable_curve(
  fit_data,
  primary_model_name,
  guess,
  known,
  upper = NULL,
  lower = NULL
)
}
\description{
Most Probable Curve method
}
