% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiple_environments}
\alias{multiple_environments}
\title{Environmental conditions for global fitting}
\format{
A list of length two, with each element containing a tibble of 3 rows and 3 columns:
\describe{
\item{time}{treatment time in min}
\item{temp}{treatment temperature (C)}
\item{pH}{treatment pH}
}
}
\usage{
multiple_environments
}
\description{
An example dataset to illustrate \code{\link[=fit_inactivation]{fit_inactivation()}} using the global approach.
It simulates the change in the environmental conditions in  2 independent experiments
at dynamic conditions. It should be used in combination with \link{multiple_counts}
}
\keyword{datasets}
