% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_step_fit.R
\name{onestep_residuals}
\alias{onestep_residuals}
\title{Residuals for one-step fitting}
\usage{
onestep_residuals(this_p, fit_data, primary_model_name, sec_models, known)
}
\description{
Residuals for one-step fitting
}
