% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InactivationPrediction_class.R
\name{plot_constant_inactivation}
\alias{plot_constant_inactivation}
\title{Plotting of inactivation under constant conditions}
\usage{
plot_constant_inactivation(
  x,
  y = NULL,
  ...,
  line_col = "black",
  line_size = 1,
  line_type = "solid",
  ylims = NULL,
  label_y = NULL,
  label_x = "time"
)
}
\description{
Plotting of inactivation under constant conditions
}
