% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InactivationPrediction_class.R
\name{plot_dynamic_inactivation}
\alias{plot_dynamic_inactivation}
\title{Plots for fynamic inactivation experiments}
\usage{
plot_dynamic_inactivation(
  x,
  add_factor = add_factor,
  ylims = ylims,
  label_y1 = label_y1,
  label_y2 = label_y2,
  line_col = line_col,
  line_size = line_size,
  line_type = line_type,
  line_col2 = line_col2,
  line_size2 = line_size2,
  line_type2 = line_type2,
  label_x = label_x,
  type = type
)
}
\description{
Plots for fynamic inactivation experiments
}
