% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivation_dynamic.R
\name{predict_dynamic_inactivation}
\alias{predict_dynamic_inactivation}
\title{Predictions under dynamic environmental conditions}
\usage{
predict_dynamic_inactivation(
  times,
  primary_model,
  env_conditions,
  secondary_models,
  ...,
  check = TRUE
)
}
\arguments{
\item{times}{a numeric vector of time points where to calculate the solution}

\item{primary_model}{a list defining the primary model as in \code{\link[=predict_inactivation]{predict_inactivation()}}}

\item{env_conditions}{a tibble (or data.frame) describing the variation of the environmental
conditions as in \code{\link[=predict_inactivation]{predict_inactivation()}}}

\item{secondary_models}{a nested list describing the secondary models for each parameter,
as defined in \link{predict_inactivation}}

\item{...}{additional arguments passed to ode}

\item{check}{whether to do some basic model checks. \code{TRUE} by default}
}
\description{
The solution is calculated by solving the differential equation using numerical
methods through the ode function from package deSolve.
}
