% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_fit.R
\name{primary_residuals}
\alias{primary_residuals}
\title{Residuals for fitting primary inactivation models}
\usage{
primary_residuals(this_p, fit_data, model_name, known)
}
\arguments{
\item{this_p}{named numeric vector with the candidate model parameters}

\item{fit_data}{data for the fit, as a tibble (or data frame) with two columns: time and logN}

\item{model_name}{model identifier}

\item{known}{named numeric vector with the known parameters}
}
\value{
an instance of modCost
}
\description{
The function is prepared to be called by modFit or modMCMC
}
