% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_models.R
\name{sec_Acclimation}
\alias{sec_Acclimation}
\title{Secondary model for p in the acclimation model}
\usage{
sec_Acclimation(x, E, Xsi)
}
\arguments{
\item{x}{value of the environmental condition}

\item{E}{E-parameter of the secondary model}

\item{Xsi}{value of x inducing stress response}
}
\description{
Secondary model for p in the acclimation model
}
