% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_models.R
\name{sec_Arrhenius}
\alias{sec_Arrhenius}
\title{Arrhenius-type secondary model}
\usage{
sec_Arrhenius(x, xref, Ea)
}
\arguments{
\item{x}{value of the environmental condition}

\item{xref}{reference value of x}

\item{Ea}{Activation energy}
}
\description{
Arrhenius-type secondary model
}
