% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_models.R
\name{sec_Bigelow}
\alias{sec_Bigelow}
\title{Bigelow-type secondary model}
\usage{
sec_Bigelow(x, xref, z)
}
\arguments{
\item{x}{value of the environmental condition}

\item{xref}{reference value of x}

\item{z}{z-value}
}
\description{
Bigelow-type secondary model
}
