% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_models.R
\name{sec_genBigelow}
\alias{sec_genBigelow}
\title{Genearlized-Bigelow secondary model}
\usage{
sec_genBigelow(x, xref, z, n)
}
\arguments{
\item{x}{value of the environmental condition}

\item{xref}{reference value of x}

\item{z}{z-value}

\item{n}{order of the model}
}
\description{
Genearlized-Bigelow secondary model
}
