% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_models.R
\name{sec_Lineal}
\alias{sec_Lineal}
\title{Lineal secondary model}
\usage{
sec_Lineal(x, xref, b)
}
\arguments{
\item{x}{value of the environmental condition}

\item{xref}{reference value of x}

\item{b}{slope of the relation}
}
\description{
Lineal secondary model
}
