% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_models.R
\name{sec_logExponential}
\alias{sec_logExponential}
\title{Log-exponential secondary model}
\usage{
sec_logExponential(x, k, xc)
}
\arguments{
\item{x}{value of the environmental condition}

\item{k}{k-parameter of the secondary model}

\item{xc}{critical value of x}
}
\description{
Log-exponential secondary model
}
