% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_guess.R
\name{show_guess_global}
\alias{show_guess_global}
\title{Showing initial guesses for global fitting}
\usage{
show_guess_global(
  fit_data,
  primary_model_name,
  guess,
  sec_models,
  env_conditions
)
}
\arguments{
\item{fit_data}{a list of tibble (or data.frame) describing the observations as
in \code{\link[=fit_inactivation]{fit_inactivation()}}}

\item{primary_model_name}{model identifier as per \code{\link[=primary_model_data]{primary_model_data()}}}

\item{guess}{a named numeric vector with the initial guess, as in \code{\link[=predict_inactivation]{predict_inactivation()}}}

\item{sec_models}{a nested list defining the secondary models as in \code{\link[=fit_inactivation]{fit_inactivation()}}}

\item{env_conditions}{a list of tibble (or data.frame) describing the environmental conditions as
in \code{\link[=fit_inactivation]{fit_inactivation()}}}
}
\value{
an instance of ggplot comparing the prediction against the data
}
\description{
Showing initial guesses for global fitting
}
