% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_guess.R
\name{show_guess_primary}
\alias{show_guess_primary}
\title{Showing initial guesses for primary models}
\usage{
show_guess_primary(fit_data, primary_model_name, guess)
}
\description{
Showing initial guesses for primary models
}
