% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_guess.R
\name{check_inactivation_guess}
\alias{check_inactivation_guess}
\title{Checking guesses for inactivation models}
\usage{
check_inactivation_guess(
  method,
  fit_data,
  primary_model_name,
  guess,
  secondary_models = NULL,
  env_conditions = NULL,
  formula = logN ~ time
)
}
\description{
Checking guesses for inactivation models
}
