% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_inactivation.R
\name{fit_inactivation}
\alias{fit_inactivation}
\title{Fits an inactivation model using different methods from predictive microbiology}
\usage{
fit_inactivation(
  method,
  fit_data,
  primary_model_name,
  guess,
  known,
  upper = NULL,
  lower = NULL,
  secondary_models = NULL,
  algorithm = "regression",
  env_conditions = NULL,
  niter = NULL,
  check = TRUE
)
}
\description{
Fits an inactivation model using different methods from predictive microbiology
}
