% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_inactivation_secondary.R
\name{fit_inactivation_secondary}
\alias{fit_inactivation_secondary}
\title{Fits a secondary inactivation model using different methods from predictive microbiology}
\usage{
fit_inactivation_secondary(
  fit_data,
  model_name,
  start,
  known,
  upper = NULL,
  lower = NULL,
  algorithm = "regression",
  niter = NULL,
  formula = my_par ~ temp,
  ...
)
}
\description{
Fits a secondary inactivation model using different methods from predictive microbiology
}
