% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_global.R
\name{fit_multiple_inactivation}
\alias{fit_multiple_inactivation}
\title{Fitting a single model to several dynamic experiments}
\usage{
fit_multiple_inactivation(
  fit_data,
  primary_model_name,
  guess,
  known,
  upper,
  lower,
  secondary_models,
  algorithm,
  env_conditions,
  niter
)
}
\description{
Fitting a single model to several dynamic experiments
}
