% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_step_fit.R
\name{fit_onestep}
\alias{fit_onestep}
\title{One-step fitting of isothermal inactivation data}
\usage{
fit_onestep(
  fit_data,
  model_name,
  start,
  known,
  upper,
  lower,
  secondary_models,
  algorithm,
  niter
)
}
\description{
One-step fitting of isothermal inactivation data
}
