% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_fit.R
\name{fit_primary}
\alias{fit_primary}
\title{Fitting of primary inactivation models}
\usage{
fit_primary(
  fit_data,
  model_name,
  start,
  known,
  upper = NULL,
  lower = NULL,
  algorithm = "regression",
  niter = NULL,
  ...,
  check = TRUE
)
}
\arguments{
\item{fit_data}{aa}

\item{model_name}{description}

\item{start}{description}

\item{known}{description}

\item{upper}{description}

\item{lower}{description}

\item{check}{description}
}
\value{
An instance of modFit with the fitted model.
}
\description{
Fitting of primary inactivation models
}
