% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_global.R
\name{get_multi_dyna_residuals}
\alias{get_multi_dyna_residuals}
\title{Residuals of multiple dynamic predictions}
\usage{
get_multi_dyna_residuals(
  this_p,
  experiment_data,
  known,
  primary_model_name,
  sec_models
)
}
\arguments{
\item{experiment_data}{a nested list with the experimental data. Each entry describes
one experiment as a list with two elements: data and conditions. \code{data} is a tibble
with two columns: time and logN. \code{conditions} is a tibble with one column named time
and as many additional columns as environmental factors.}
}
\value{
an instance of `modCost.
}
\description{
Residuals of multiple dynamic predictions
}
