% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_guesses.R
\name{make_guess_secondary}
\alias{make_guess_secondary}
\title{Initial guesses for fitting secondary inactivation models}
\usage{
make_guess_secondary(fit_data, model_name, formula = my_par ~ temp)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The function uses some heuristics to provide initial guesses for the parameters
of the growth model selected that can be used with \code{\link[=fit_inactivation_secondary]{fit_inactivation_secondary()}}.
}
