% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivation_dynamic.R
\name{predict_dynamic_inactivation}
\alias{predict_dynamic_inactivation}
\title{Predictions under dynamic environmental conditions}
\usage{
predict_dynamic_inactivation(
  times,
  primary_model,
  env_conditions,
  secondary_models,
  check = TRUE
)
}
\description{
Predictions under dynamic environmental conditions
}
