% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_inactivation_uncertainty.R
\name{predict_inactivation_uncertainty}
\alias{predict_inactivation_uncertainty}
\title{AA}
\usage{
predict_inactivation_uncertainty(
  model_name,
  times,
  n_sims,
  pars,
  corr_matrix = diag(nrow(pars))
)
}
\description{
AA
}
