% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecondaryFit_class.R
\name{SecondaryFit}
\alias{SecondaryFit}
\alias{coef.SecondaryFit}
\alias{summary.SecondaryFit}
\title{SecondaryFit class}
\usage{
\method{coef}{SecondaryFit}(object, ...)

\method{summary}{SecondaryFit}(object, ...)
}
\arguments{
\item{object}{Instance of \link{SecondaryFit}}

\item{...}{ignored}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The \code{SecondaryFit} class contains a secondary inactivation model fitted to a set of
parameters obtained under static environmental conditions.
Its constructor is \code{\link[=fit_inactivation_secondary]{fit_inactivation_secondary()}}.

It is a subclass of list with the items:
\itemize{
\item ...:...
}
}
\section{Methods (by generic)}{
\itemize{
\item \code{coef(SecondaryFit)}: vector of fitted model parameters.

\item \code{summary(SecondaryFit)}: statistical summary of the fit.

}}
